requirements.txt
carvekit_colab==4.1.0
dl_ext==1.3.4
einops==0.3.0
imageio==2.9.0
iopath==0.1.10
ipython==8.12.3
kornia==0.6.0
loguru==0.7.0
matplotlib==3.7.2
numpy==1.24.4
omegaconf==2.1.1
opencv_python==4.5.5.64
opencv_python_headless==4.8.0.74
pandas==2.0.3
Pillow==9.4.0
plotly==5.13.1
psutil==5.9.7
pudb==2019.2
pytorch_lightning==1.4.2
pytz==2023.3
PyYAML==6.0.1
scipy==1.9.1
six==1.16.0
tabulate==0.9.0
torch==1.12.1+cu116
torchvision==0.13.1+cu116
tqdm==4.65.0
transformers==4.22.2
transforms3d==0.4.1
webdataset==0.2.5
yacs==0.1.8
